# swagger-spec.yml
info:
  title: "Node Express Server"
  version: "1.0.0"
  description: "Node Express Server With User Creation, Authentication, Password Reset using Email Verification and JWT's."
openapi: "3.0.0"
paths:
...
  "/journal/edit":
    post:
      description: "Edits a journals title, entry and/or category."
      tags:
        - Journal
      requestBody:
        description: "The title, entry or category is required."
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                journalId:
                  type: string
                title:
                  type: string
                  nullable: true
                entry:
                  type: string
                  nullable: true
                category:
                  type: string
                  nullable: true
            example:
              userId: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9"
              journalId: "eyJhbGciOiJIUzI1NiI4234242423423VCJ9"
              title: "New Title"
              entry: "New entry"
              category: "New Category"
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    allOf:
                      - $ref: "#/components/schemas/User"
                      - nullable: true
              example:
                success: true
                message: "Journal edited."
                data:
                  {
                    "username": "user1",
                    "email": "email@gmail.com",
                    "journals": [],
                    "journalCategories": [],
                  }
        "422":
          description: "Invalid request body"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    nullable: true
              example:
                success: false
                message: "Invalid request body."
                data: null
        "500":
          description: "Caught error"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    nullable: true
              example:
                success: false
                message: "An error occurred. Error: something went wrong."
                data: null
  "/journal/journals/{userId}":
    get:
      summary: "Get a users journals."
      tags:
        - Journal
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
          description: "The users document ID"
          example: "666655343535a02523522ac32423"
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    allOf:
                      - $ref: "#/components/schemas/User"
              example:
                success: true
                message: "Successfully fetched the journals."
                data:
                  {
                    "username": "user1",
                    "email": "email@gmail.com",
                    "journals": [],
                    "journalCategories": [],
                  }
        "422":
          description: "Invalid request body"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    nullable: true
              example:
                success: false
                message: "Invalid request body."
                data: null
        "500":
          description: "Caught error"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    nullable: true
              example:
                success: false
                message: "An error occurred. Error: something went wrong."
                data: null
...
  "/journal/new-category":
    post:
      description: "Saves a new category."
      tags:
        - Journal
      requestBody:
        description: "A users document ID and a category string."
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  description: "The user's document id."
                  type: string
                category:
                  description: "The category string."
                  type: string
            example:
              userId: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9"
              category: "New Category"
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    allOf:
                      - $ref: "#/components/schemas/User"
                      - nullable: true
              example:
                success: true
                message: "Category saved."
                data:
                  {
                    "username": "user1",
                    "email": "email@gmail.com",
                    "journals": [],
                    "journalCategories": ["New Category"],
                  }
        "422":
          description: "Invalid request body"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    nullable: true
              example:
                success: false
                message: "Invalid request body."
                data: null
        "500":
          description: "Caught error"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    nullable: true
              example:
                success: false
                message: "An error occurred. Error: something went wrong."
                data: null
  "/journal/bulk-set-category":
    post:
      description: "Saves a category on the journal(s)."
      tags:
        - Journal
      requestBody:
        description: "A users document ID, the journal document id's, and a category string."
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  description: "The user's document id."
                  type: string
                journalIds:
                  description: "The journal document id(s)."
                  type: string
                category:
                  description: "The category string."
                  type: string
            example:
              userId: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9"
              journalIds:
                [
                  "eyJhbGciOifgIUzI1NiIsInR5cCI6IkpXVCJ9",
                  "eyJhbhgiOiJIUzI1NiIsInR5cCI6IkpXVCJ9",
                ]
              category: "New Category"
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    allOf:
                      - $ref: "#/components/schemas/User"
                      - nullable: true
              example:
                success: true
                message: "Journals updated."
                data:
                  {
                    "username": "user1",
                    "email": "email@gmail.com",
                    "journals": [],
                    "journalCategories": ["New Category"],
                  }
        "422":
          description: "Invalid request body"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    nullable: true
              example:
                success: false
                message: "Invalid request body."
                data: null
        "500":
          description: "Caught error"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    nullable: true
              example:
                success: false
                message: "An error occurred. Error: something went wrong."
                data: null
components:
  schemas:
...
    User:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        journals:
          type: array
          items:
            $ref: "#/components/schemas/Journal"
        journalCategories:
          type: array
          items:
            type: object
            properties:
              category:
                type: string
              selected:
                type: boolean
    Journal:
      type: object
      properties:
        title:
          type: string
          example: "Journal 1"
        entry:
          type: string
          example: "Journal 1"
        category:
          type: string
        date:
          type: string
        selected:
          type: boolean
...